# apps/web/openapi.json
{
  "components": {
    "schemas": {
...
      "AgentSection": {
        "additionalProperties": false,
        "properties": {
          "allowed_tools": {
            "items": {
              "type": "string"
            },
            "title": "Allowed Tools",
            "type": "array"
          },
          "execution_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Id"
          },
          "independent": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Independent"
          },
          "input_artifact_digest": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Artifact Digest"
          },
          "instruction_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instruction Version"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "output_artifact_digest": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Artifact Digest"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/AgentRole"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "validation_evidence_set_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validation Evidence Set Id"
          }
        },
        "required": [
          "role",
          "provider",
          "model",
          "execution_id",
          "status",
          "instruction_version",
          "input_artifact_digest",
          "output_artifact_digest",
          "validation_evidence_set_id",
          "independent",
          "allowed_tools"
        ],
        "title": "AgentSection",
        "type": "object"
      },
...
      "ArtifactLineageResponse": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "parent_digests": {
            "items": {
              "type": "string"
            },
            "title": "Parent Digests",
            "type": "array"
          },
          "producer_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Producer Id"
          },
          "producer_type": {
            "title": "Producer Type",
            "type": "string"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "producer_type",
          "producer_id",
          "created_at",
          "parent_digests"
        ],
        "title": "ArtifactLineageResponse",
        "type": "object"
      },
      "ArtifactResponse": {
        "additionalProperties": false,
        "properties": {
          "byte_count": {
            "title": "Byte Count",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "digest": {
            "title": "Digest",
            "type": "string"
          },
          "lineage": {
            "items": {
              "$ref": "#/components/schemas/ArtifactLineageResponse"
            },
            "title": "Lineage",
            "type": "array"
          },
          "media_type": {
            "title": "Media Type",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "original_byte_count": {
            "title": "Original Byte Count",
            "type": "integer"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "integer"
          },
          "truncated": {
            "title": "Truncated",
            "type": "boolean"
          },
          "truncation_policy": {
            "title": "Truncation Policy",
            "type": "string"
          }
        },
        "required": [
          "digest",
          "media_type",
          "byte_count",
          "schema_version",
          "created_at",
          "metadata",
          "truncated",
          "original_byte_count",
          "truncation_policy",
          "lineage"
        ],
        "title": "ArtifactResponse",
        "type": "object"
      },
      "ArtifactTextResponse": {
        "additionalProperties": false,
        "properties": {
          "digest": {
            "title": "Digest",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "digest",
          "text"
        ],
        "title": "ArtifactTextResponse",
        "type": "object"
      },
...
      "AvailableCommand": {
        "additionalProperties": false,
        "properties": {
          "evidence_digest": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Digest"
          },
          "expected_run_version": {
            "title": "Expected Run Version",
            "type": "integer"
          },
          "gate": {
            "anyOf": [
              {
                "enum": [
                  "plan",
                  "pr",
                  "merge"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gate"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "policy_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy Version"
          },
          "requires_feedback": {
            "title": "Requires Feedback",
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "expected_run_version",
          "requires_feedback"
        ],
        "title": "AvailableCommand",
        "type": "object"
      },
...
      "BudgetSection": {
        "additionalProperties": false,
        "properties": {
          "cost_limit_minor": {
            "title": "Cost Limit Minor",
            "type": "integer"
          },
          "duration_limit_seconds": {
            "title": "Duration Limit Seconds",
            "type": "integer"
          },
          "local_remediation_count": {
            "title": "Local Remediation Count",
            "type": "integer"
          },
          "local_remediation_limit": {
            "title": "Local Remediation Limit",
            "type": "integer"
          },
          "local_remediation_remaining": {
            "title": "Local Remediation Remaining",
            "type": "integer"
          },
          "remote_remediation_count": {
            "title": "Remote Remediation Count",
            "type": "integer"
          },
          "remote_remediation_limit": {
            "title": "Remote Remediation Limit",
            "type": "integer"
          },
          "remote_remediation_remaining": {
            "title": "Remote Remediation Remaining",
            "type": "integer"
          },
          "token_limit": {
            "title": "Token Limit",
            "type": "integer"
          }
        },
        "required": [
          "local_remediation_count",
          "local_remediation_limit",
          "local_remediation_remaining",
          "remote_remediation_count",
          "remote_remediation_limit",
          "remote_remediation_remaining",
          "token_limit",
          "cost_limit_minor",
          "duration_limit_seconds"
        ],
        "title": "BudgetSection",
        "type": "object"
      },
      "CandidateSection": {
        "additionalProperties": false,
        "properties": {
          "commit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commit"
          },
          "pending_evidence_digest": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pending Evidence Digest"
          },
          "review_evidence_digest": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Evidence Digest"
          },
          "validation_evidence_digest": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validation Evidence Digest"
          }
        },
        "required": [
          "commit",
          "pending_evidence_digest",
          "validation_evidence_digest",
          "review_evidence_digest"
        ],
        "title": "CandidateSection",
        "type": "object"
      },
      "CheckItem": {
        "additionalProperties": false,
        "properties": {
          "command_name": {
            "title": "Command Name",
            "type": "string"
          },
          "command_version": {
            "title": "Command Version",
            "type": "integer"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "exit_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exit Code"
          },
          "head_sha": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Head Sha"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "output_artifact_digest": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Artifact Digest"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "command_name",
          "command_version",
          "status",
          "exit_code",
          "output_artifact_digest",
          "completed_at",
          "head_sha"
        ],
        "title": "CheckItem",
        "type": "object"
      },
...
      "DashboardSummary": {
        "additionalProperties": false,
        "properties": {
          "runs": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Runs",
            "type": "object"
          },
          "total_runs": {
            "title": "Total Runs",
            "type": "integer"
          }
        },
        "required": [
          "total_runs",
          "runs"
        ],
        "title": "DashboardSummary",
        "type": "object"
      },
...
      "EventItem": {
        "additionalProperties": false,
        "properties": {
          "actor_class": {
            "title": "Actor Class",
            "type": "string"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "run_version": {
            "title": "Run Version",
            "type": "integer"
          },
          "sequence": {
            "title": "Sequence",
            "type": "integer"
          }
        },
        "required": [
          "sequence",
          "event_type",
          "run_version",
          "actor_class",
          "occurred_at",
          "payload"
        ],
        "title": "EventItem",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlanSection": {
        "additionalProperties": false,
        "properties": {
          "approval_evidence_digest": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Evidence Digest"
          },
          "approval_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Id"
          },
          "output_artifact_digest": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Artifact Digest"
          }
        },
        "required": [
          "output_artifact_digest",
          "approval_id",
          "approval_evidence_digest"
        ],
        "title": "PlanSection",
        "type": "object"
      },
...
      "ProjectSection": {
        "additionalProperties": false,
        "properties": {
          "github_repository": {
            "title": "Github Repository",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "policy_digest": {
            "title": "Policy Digest",
            "type": "string"
          },
          "policy_version": {
            "title": "Policy Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "github_repository",
          "policy_version",
          "policy_digest"
        ],
        "title": "ProjectSection",
        "type": "object"
      },
      "PullRequestSection": {
        "additionalProperties": false,
        "properties": {
          "base_ref": {
            "title": "Base Ref",
            "type": "string"
          },
          "base_sha": {
            "title": "Base Sha",
            "type": "string"
          },
          "branch": {
            "title": "Branch",
            "type": "string"
          },
          "head_sha": {
            "title": "Head Sha",
            "type": "string"
          },
          "merge_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Merge State"
          },
          "number": {
            "title": "Number",
            "type": "integer"
          },
          "repository": {
            "title": "Repository",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "number",
          "repository",
          "branch",
          "base_ref",
          "head_sha",
          "base_sha",
          "state",
          "merge_state"
        ],
        "title": "PullRequestSection",
        "type": "object"
      },
      "ResourceSection": {
        "additionalProperties": false,
        "properties": {
          "branch_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Branch Name"
          },
          "database_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Database Name"
          },
          "database_state": {
            "$ref": "#/components/schemas/ResourceState"
          },
          "worktree_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Worktree Path"
          }
        },
        "required": [
          "worktree_path",
          "branch_name",
          "database_state",
          "database_name"
        ],
        "title": "ResourceSection",
        "type": "object"
      },
...
      "RunProjection": {
        "additionalProperties": false,
        "properties": {
          "agents": {
            "additionalProperties": {
              "$ref": "#/components/schemas/AgentSection"
            },
            "title": "Agents",
            "type": "object"
          },
          "available_commands": {
            "items": {
              "$ref": "#/components/schemas/AvailableCommand"
            },
            "title": "Available Commands",
            "type": "array"
          },
          "budgets": {
            "$ref": "#/components/schemas/BudgetSection"
          },
          "candidate": {
            "$ref": "#/components/schemas/CandidateSection"
          },
          "checks": {
            "items": {
              "$ref": "#/components/schemas/CheckItem"
            },
            "title": "Checks",
            "type": "array"
          },
          "latest_events": {
            "items": {
              "$ref": "#/components/schemas/EventItem"
            },
            "title": "Latest Events",
            "type": "array"
          },
          "next_gate": {
            "anyOf": [
              {
                "enum": [
                  "plan",
                  "pr",
                  "merge"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Gate"
          },
          "plan": {
            "$ref": "#/components/schemas/PlanSection"
          },
          "project": {
            "$ref": "#/components/schemas/ProjectSection"
          },
          "pull_request": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PullRequestSection"
              },
              {
                "type": "null"
              }
            ]
          },
          "resource": {
            "$ref": "#/components/schemas/ResourceSection"
          },
          "review": {
            "$ref": "#/components/schemas/ReviewSection"
          },
          "run": {
            "$ref": "#/components/schemas/RunResponse"
          },
          "security": {
            "$ref": "#/components/schemas/SecuritySection"
          },
          "task": {
            "$ref": "#/components/schemas/TaskSection"
          },
          "usage": {
            "$ref": "#/components/schemas/UsageSummary"
          }
        },
        "required": [
          "run",
          "task",
          "project",
          "resource",
          "plan",
          "candidate",
          "pull_request",
          "checks",
          "review",
          "agents",
          "budgets",
          "usage",
          "security",
          "latest_events",
          "available_commands",
          "next_gate"
        ],
        "title": "RunProjection",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "CSRF": {
        "in": "header",
        "name": "X-CSRF-Token",
        "type": "apiKey"
      },
      "OperatorSession": {
        "in": "cookie",
        "name": "forge_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Parallel Forge",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/artifacts/{digest}": {
      "get": {
        "operationId": "get_artifact_api_artifacts__digest__get",
        "parameters": [
          {
            "in": "path",
            "name": "digest",
            "required": true,
            "schema": {
              "title": "Digest",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OperatorSession": []
          }
        ],
        "summary": "Get Artifact"
      }
    },
    "/api/artifacts/{digest}/download": {
      "get": {
        "operationId": "download_api_artifacts__digest__download_get",
        "parameters": [
          {
            "in": "path",
            "name": "digest",
            "required": true,
            "schema": {
              "title": "Digest",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OperatorSession": []
          }
        ],
        "summary": "Download"
      }
    },
    "/api/artifacts/{digest}/text": {
      "get": {
        "operationId": "get_text_api_artifacts__digest__text_get",
        "parameters": [
          {
            "in": "path",
            "name": "digest",
            "required": true,
            "schema": {
              "title": "Digest",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactTextResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OperatorSession": []
          }
        ],
        "summary": "Get Text"
      }
    },
...
    "/api/dashboard/summary": {
      "get": {
        "operationId": "summary_api_dashboard_summary_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardSummary"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OperatorSession": []
          }
        ],
        "summary": "Summary"
      }
    },
...
    "/api/runs/{run_id}/projection": {
      "get": {
        "operationId": "projection_api_runs__run_id__projection_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunProjection"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OperatorSession": []
          }
        ],
        "summary": "Projection"
      }
    },
...
  }
}